+++ src/creatures/players/player.cpp
@@ -167,6 +167,8 @@ std::string Player::getDescription(int32_t lookDistance) {
 
 		if (group->access) {
 			s << " You are " << group->name << '.';
+		} else if (hasFlag(PlayerFlags_t::IsGameTester)) {
+			s << " You are a " << group->name << '.';
 		} else if (vocation->getId() != VOCATION_NONE) {
 			s << " You are " << vocation->getVocDescription() << '.';
 		} else {
@@ -190,6 +192,8 @@ std::string Player::getDescription(int32_t lookDistance) {
 
 		if (group->access) {
 			s << " " << getSubjectVerb() << " " << group->name << '.';
+		} else if (hasFlag(PlayerFlags_t::IsGameTester)) {
+			s << " " << getSubjectVerb() << " a " << group->name << '.';
 		} else if (vocation->getId() != VOCATION_NONE) {
 			s << " " << getSubjectVerb() << " " << vocation->getVocDescription() << '.';
 		} else {
@@ -3387,7 +3391,7 @@ BlockType_t Player::blockHit(const std::shared_ptr<Creature> &attacker, const Co
 				}
 			}
 
-			//
+			// Absorb Percent
 			const ItemType &it = Item::items[item->getID()];
 			if (it.abilities) {
 				int totalAbsorbPercent = 0;
@@ -3403,7 +3407,7 @@ BlockType_t Player::blockHit(const std::shared_ptr<Creature> &attacker, const Co
 					}
 				}
 
-				if (totalAbsorbPercent > 0) {
+				if (totalAbsorbPercent != 0) {
 					damage -= std::round(damage * (totalAbsorbPercent / 100.0));
 
 					const auto charges = item->getAttribute<uint16_t>(ItemAttribute_t::CHARGES);
@@ -6126,7 +6130,7 @@ bool Player::lastHitIsPlayer(const std::shared_ptr<Creature> &lastHitCreature) {
 }
 
 void Player::changeHealth(int32_t healthChange, bool sendHealthChange /* = true*/) {
-	if (g_configManager().getBoolean(TOGGLE_TEST_MODE)) {
+	if (hasFlag(PlayerFlags_t::IsGameTester)) {
 		return;
 	}
 
@@ -6135,7 +6139,7 @@ void Player::changeHealth(int32_t healthChange, bool sendHealthChange /* = true*
 }
 
 void Player::changeMana(int32_t manaChange) {
-	if (g_configManager().getBoolean(TOGGLE_TEST_MODE)) {
+	if (hasFlag(PlayerFlags_t::IsGameTester)) {
 		return;
 	}
 
@@ -6147,7 +6151,7 @@ void Player::changeMana(int32_t manaChange) {
 }
 
 void Player::changeSoul(int32_t soulChange) {
-	if (g_configManager().getBoolean(TOGGLE_TEST_MODE)) {
+	if (hasFlag(PlayerFlags_t::IsGameTester)) {
 		return;
 	}
 
@@ -7214,9 +7218,16 @@ uint8_t Player::getRandomMountId() const {
 		}
 	}
 
-	const auto playerMountsSize = static_cast<int32_t>(playerMounts.size() - 1);
-	const auto randomIndex = uniform_random(0, std::max<int32_t>(0, playerMountsSize));
-	return playerMounts.at(randomIndex);
+	if (playerMounts.empty()) {
+		return 0;
+	}
+
+	const auto randomIndex = uniform_random(0, static_cast<int32_t>(playerMounts.size() - 1));
+	if (randomIndex >= 0 && static_cast<size_t>(randomIndex) < playerMounts.size()) {
+		return playerMounts[randomIndex];
+	}
+
+	return 0;
 }
 
 bool Player::toggleMount(bool mount) {
@@ -10269,7 +10280,6 @@ void Player::onFollowCreatureDisappear(bool isLogout) {
 	}
 }
 
-// container
 // container
 
 void Player::onAddContainerItem(const std::shared_ptr<Item> &item) {