+++ src/lua/creature/movement.cpp
@@ -816,6 +816,13 @@ uint32_t MoveEvent::fireAddRemItem(const std::shared_ptr<Item> &item, const std:
 	if (isLoadedScriptId()) {
 		return executeAddRemItem(item, fromTile, pos);
 	} else {
+		if (!moveFunction) {
+			g_logger().error("[MoveEvent::fireAddRemItem - Item {} item on position: {}] "
+			                 "Move function is nullptr.",
+			                 item->getName(), pos.toString());
+			return 0;
+		}
+
 		return moveFunction(item, fromTile, pos);
 	}
 }
@@ -848,6 +855,13 @@ uint32_t MoveEvent::fireAddRemItem(const std::shared_ptr<Item> &item, const Posi
 	if (isLoadedScriptId()) {
 		return executeAddRemItem(item, pos);
 	} else {
+		if (!moveFunction) {
+			g_logger().error("[MoveEvent::fireAddRemItem - Item {} item on position: {}] "
+			                 "Move function is nullptr.",
+			                 item->getName(), pos.toString());
+			return 0;
+		}
+
 		return moveFunction(item, nullptr, pos);
 	}
 }
@@ -857,9 +871,9 @@ bool MoveEvent::executeAddRemItem(const std::shared_ptr<Item> &item, const Posit
 	// onRemoveItem(moveitem, pos)
 	if (!LuaScriptInterface::reserveScriptEnv()) {
 		g_logger().error("[MoveEvent::executeAddRemItem - "
-		                 "Item {} item on tile x: {} y: {} z: {}] "
+		                 "Item {} item on position: {}] "
 		                 "Call stack overflow. Too many lua script calls being nested.",
-		                 item->getName(), pos.getX(), pos.getY(), pos.getZ());
+		                 item->getName(), pos.toString());
 		return false;
 	}